requirements.txt
alembic==1.18.1
APScheduler==3.11.2
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==24.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
python-dotenv==1.2.1
PyYAML==6.0.3
razorpay==2.0.0
requests==2.32.5
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==202601
urllib3==2.6.3
user-agents==2.2.0
waitress==3.0.2
Werkzeug==3.1.5
